requirements.txt
pika==1.3.2
sqlalchemy==2.0.43
python-dotenv==1.1.1
pydantic==2.10.6
psycopg2-binary==2.9.10
requests==2.32.3
fastapi==0.116.1
uvicorn==0.29.0
pillow==10.4.0
python-multipart==0.0.9
pytest==8.4.1
httpx==0.28.1
google-auth==2.28.1

opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-jaeger==1.21.0

opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-pika==0.59b0

deprecated==1.2.18
python-json-logger==4.0.0

prometheus-fastapi-instrumentator==7.1.0

black==25.9.0
isort==7.0.0
pydantic-settings==2.11.0